refactor: move filters to where they belong
