+++ src/sandbox/services/filters/index.js
@@ -1,0 +1,4 @@
+import Vue from 'vue';
+import { formatCode } from '@/sandbox/services/filters/formatCode';
+
+Vue.filter('formatCode', formatCode);
